fix: fix wrong return on jwt reset strategy
