+++ backend/src/modules/auth/services/password-recovery.service.ts
@@ -1,28 +1,25 @@
 import { Inject, Injectable, Logger } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { JwtService } from '@nestjs/jwt';
 import * as bcrypt from 'bcrypt';
 import { VERIFICATION } from 'src/common/constants/verification.constants';
 import { IUserService, USER_SERVICE } from 'src/modules/users/interfaces/user-service.interface';
 import { IVerificationService } from 'src/modules/verification/interfaces/verification-service.interface';
-import { JwtConfig, jwtConfig } from '../config/jwt.config';
 import { InvalidPasswordResetCodeException } from '../exceptions/auth.exceptions';
 import { IPasswordRecoveryService } from '../interfaces/password-recovery-service.interface';
+import { TokenFactoryService } from './token-factory.service';
+import { User } from 'src/database/entities/user.entity';
 
 @Injectable()
 export class PasswordRecoveryService implements IPasswordRecoveryService {
     private readonly logger = new Logger(PasswordRecoveryService.name);
-    private readonly jwtConfig: JwtConfig;
 
     constructor(
         @Inject(USER_SERVICE) private readonly usersService: IUserService,
         @Inject(VERIFICATION.RECOVER_PASSWORD_TOKEN)
         private readonly verificationService: IVerificationService,
-        private readonly jwtService: JwtService,
+        private readonly tokenFactory: TokenFactoryService,
         private readonly configService: ConfigService,
-    ) {
-        this.jwtConfig = jwtConfig(this.configService);
-    }
+    ) { }
 
     async sendResetPasswordCode(phoneNumber: string): Promise<void> {
         try {
@@ -53,17 +50,9 @@ export class PasswordRecoveryService implements IPasswordRecoveryService {
 
         try {
             const user = await this.usersService.findByPhoneNumber(phoneNumber);
-
-            const payload = {
-                sub: user.id,
-                scope: 'reset_password' as const,
-            };
-
-            const token = await this.jwtService.signAsync(payload, {
-                expiresIn: this.jwtConfig.resetTokenExpiration,
-                secret: this.jwtConfig.resetSecret,
-            });
-
+            this.logger.log(`Generating reset password token for user: ${user?.id}`);
+            const token = await this.tokenFactory.generateResetPasswordToken(user as User); 
+            this.logger.log(`Reset password token generated for ${phoneNumber}`);
             return {
                 token,
                 message: 'Password reset token generated successfully',
@@ -78,12 +67,14 @@ export class PasswordRecoveryService implements IPasswordRecoveryService {
 
     async resetPassword(id: number, newPassword: string): Promise<string> {
         try {
+            this.logger.log(`Resetting password for user ID: ${id}`);
             const user = await this.usersService.findByIdWithPassword(id);
+            this.logger.log(`User found: ${user?.id}`);
             const newHashedPassword = await bcrypt.hash(newPassword, 10);
-
+            this.logger.log(`New password hashed for user ID: ${id}`);
             user.password = newHashedPassword;
             user.passwordUpdatedAt = new Date();
-
+            this.logger.log(`Updating user password in database for user ID: ${id}`);
             await this.usersService.update(user.id, user);
             return 'Password reset successful';
         } catch (error) {
